refactor: Update PushSubscription model
This commit updates the PushSubscription model in the Back-end/src/main/java/com/CondoSync/models/PushSubscription.java file. It adds a new column `hashKey` and removes the `setHashKey` method. The `hashKey` column is now nullable and has a unique constraint. This change improves the data structure and ensures uniqueness of the `hashKey` values.