feat: Add SubscripitionDTO class
This commit adds the SubscripitionDTO class to the Back-end/src/main/java/com/CondoSync/models/DTOs directory. The SubscripitionDTO class is a data transfer object that represents a subscription for push notifications. It includes fields for the endpoint, expiration time, and keys. The keys field is a nested class that contains the p256dh and auth keys. This class will be used to handle subscription data in the application.