refactor: Update MoradorService to handle morador-related operations
This commit adds the MoradorService class to the Back-end/src/main/java/com/CondoSync/services directory. The MoradorService class provides methods for finding a morador by CPF or email, as well as saving a new morador. It uses the MoradorRepository to interact with the database. This service will be used to handle morador-related operations in the application.