refactor: Update UserService to include password encoding
This commit adds the PasswordEncoder dependency to the UserService class in the Back-end/src/main/java/com/CondoSync/services/UserService.java file. The PasswordEncoder will be used to encode passwords before storing them in the database. This change improves security by ensuring that passwords are not stored in plain text.