feat: Add user registration endpoint to MoradorController
This commit adds a new endpoint `/register` to the `MoradorController` class in the `Back-end/src/main/java/com/CondoSync/controllers/MoradorController.java` file. The endpoint handles the registration of a new user by accepting a `MoradorDTO` object in the request body. The `MoradorService` is used to register the user, and if successful, the created `MoradorDTO` is returned with a `201 Created` status. If the registration fails, a `400 Bad Request` status is returned. This addition enhances the functionality of the application by allowing users to register and join the system.