feat: Update Area model with price validation
The `Area` model in the `Area.java` file has been updated to include price validation. The `price` field now has a `@NotNull` annotation and a `@DecimalMin` annotation to ensure that the price is not null and greater than zero. This change improves the data integrity and validation of the `Area` model.