feat: Add register and listAll methods to AreaService
The `AreaService` class in the `AreaService.java` file has been updated with the addition of the `register` and `listAll` methods. The `register` method takes a `@Valid` `AreaDTO` as input and creates a new `Area` entity with the provided data, including images. The `listAll` method retrieves all areas from the repository and converts them into a list of `AreaDTO` objects. These changes enhance the functionality of the `AreaService` and improve the data handling capabilities of the application.