feat: Add Horario model for scheduling
A new model class `Horario` has been added in the `Horario.java` file. This model represents a schedule with `horaInicio` and `horaFim` fields of type `LocalTime`. It also has a many-to-one relationship with the `Area` entity. This addition enhances the application's data handling capabilities and allows for scheduling and managing time slots for areas.