feat: Add Horario model and update Area model with horarios field
The `Horario` model has been added in the `Horario.java` file to represent a schedule with `horaInicio` and `horaFim` fields of type `LocalTime`. It also has a many-to-one relationship with the `Area` entity. This addition enhances the application's data handling capabilities and allows for scheduling and managing time slots for areas.

The `Area` model in the `Area.java` file has been updated with the addition of the `horarios` field, which is a list of `Horario` objects. This change improves the data representation and functionality of the `Area` model, allowing for the association of multiple time slots with an area.