feat: Add findMoradorByEmail method to MoradorService
The MoradorService class in the MoradorService.java file has been updated with the addition of the findMoradorByEmail method. This method retrieves a Morador entity from the repository based on the provided email. If no Morador is found with the given email, an EntityNotFoundException is thrown. This change enhances the functionality of the MoradorService and improves the data retrieval capabilities of the application.