feat: Add Reserva model for managing reservations
A new model class `Reserva` has been added in the `Reserva.java` file. This model represents a reservation with `data`, `horaInicio`, `horaFim`, `statusReserva`, `creation`, and `update` fields. It also has a unique constraint on the combination of `data_reserva`, `hora_inicio`, and `hora_fim` columns in the database. This addition enhances the application's data handling capabilities and allows for managing reservations for areas.