refactor: Update TokenService to use List instead of String for scope
The TokenService class in TokenService.java has been updated to use a List<String> instead of a String for the scope variable. This change improves the data handling capabilities and ensures that multiple roles can be assigned to a user. The scope is now stored as a list of role authorities in the JWT claims.