refactor: Update HorarioRepository to add query for finding available horarios
The code changes in HorarioRepository.java add a new query method, findAvailableHorariosByAreaIdAndDate, to retrieve available horarios based on the areaId and date parameters. This refactor enhances the functionality of the repository by providing a convenient way to fetch horarios that are not already reserved for a specific area and date combination.