refactor: Add List<OcorrenciaMoradore> field to Morador model
The code changes in Morador.java add a new field, List<OcorrenciaMoradore> ocorrencias, to the Morador model. This refactor enhances the codebase by allowing the Morador entity to have a collection of OcorrenciaMoradore objects associated with it. This change improves the flexibility and functionality of the application by enabling the tracking and management of occurrences related to a resident.