refactor: Add UserUpdatePasswordDTO for updating user password
The code changes introduce a new class, UserUpdatePasswordDTO, in the DTOs package. This class is used for updating the user password and includes fields for the current password, new password, and confirmation of the new password. This refactor improves the code structure and separation of concerns by introducing a dedicated DTO class for updating the user password.